cv_scoring_agent/src/app/requirements.txt
python-dotenv
openpyxl
nltk
spacy
fuzzywuzzy
python-dotenv
PyPDF2
python-docx
scikit-learn
pandas
streamlit
python-dotenv
